Add files via upload
On the basis of Julie's program, I have added an undo feature and a multi-game function currently limited to player-versus-player mode. However, there is still a minor bug in the scoreboard display. I will find time to fix the bug later and add a multi-game function for AI versus AI mode.